+++ src/main/java/tn/esprit/pi/entities/Driver.java
@@ -38,7 +38,7 @@ public class Driver implements Serializable{
 		super();
 	}
 
-	public int getIdDriver() {
+	public long getIdDriver() {
 		return idDriver;
 	}
