+++ src/main/java/tn/esprit/pi/services/IBusService.java
@@ -13,10 +13,10 @@ import tn.esprit.pi.entities.Bus;
 public interface IBusService {
 	List<Bus> getAllBus();
 	
-	Bus  AffectBusDriver(Bus bus,long idDriver,int idU);
-	Bus getBusById(long id);
-	void deleteBusById(long id);
-	Bus updateBus(Bus bus, long id);
+	Bus  AffectBusDriver(Bus bus,int idDriver,int idU);
+	Bus getBusById(int id);
+	void deleteBusById(int id);
+	Bus updateBus(Bus bus, int id);
 	List<Bus> getBusByDriver(String D);
 	
 }
